add SSH to GitHub check and a note about it in the README (#48)
closes #43

According to GitHub docs, ssh -T git@github.com seems to be the proper way to test if the current user is able to connect to GitHub using SSH authentication, which at the moment is a requirement for this script to work. After some experimentation, I determined that ssh -T git@github.com will return 1 if it's successful, and probably some other positive integer if it is not.

Based on that, I added a test early on in the script to fail early with a message if the current user cannot connect to GitHub with SSH, and I also added a note to the README that SSH is necessary to use this tool.